Refactor technology handling to use ScreeningTechnology
Transitioned from HallTechnology to ScreeningTechnology across the application. Updated constructors, properties, and methods in `CinemaAppService.cs`, `CinemaHall`, and `Movie` classes. Modified `MovieCreatedEvent` to include ScreeningTechnology. Introduced `CinemaHallCreatedEventHandler` and `MovieCreatedEventHandler` for event handling. Updated database schema to accommodate new data types and added SeatCount. Enhanced integration event bus with cancellation token support. Overall, these changes improve the flexibility and functionality of the cinema management system.

+++ CinemaTicketingSystem.Domain/CinemaManagement/Seat.cs
@@ -17,13 +17,13 @@ public class Seat : Entity<Guid>
             throw new ArgumentException("Seat number must be positive");
 
         Row = row.ToUpper();
-        Number = number;
+        Number = (short)number;
         Type = type;
         Id = Guid.CreateVersion7();
     }
 
     public string Row { get; } = null!;
-    public int Number { get; }
+    public short Number { get; }
     public SeatType Type { get; private set; }
     public bool IsAvailable { get; private set; } = true;
